refactor: move useEncryptionKey hook to its own file
