+++ frontend/utils/useEncryptionKey.ts
